Created AppLayout with Header and Footer
