+++ app/(routes)/dashboard/DashboardPage.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import AppLayout from "@/app/_common/layout/AppLayout";
+import DashboardContent from "./DashboardContent";
+
+const DashboardPage = () => {
+  return (
+    <AppLayout>
+      <DashboardContent />
+    </AppLayout>
+  );
+};
+
+export default DashboardPage;
