+++ app/_common/layout/UsFooter.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+const UsFooter = () => {
+
+  return (
+    <div className="h-10 bg-gray-200 p-4 flex items-center justify-center text-sm">
+      Footer
+    </div>
+  );
+};
+
+export default UsFooter;
