+++ app/_utilities/GlobalVariables.ts
@@ -1,0 +1,3 @@
+export const oneToOneHundredSelectOptions = Array.from({ length: 100 }, (_, i) => ({
+  value: (i + 1).toString(),
+}));
